+++ index.js
@@ -1,4 +1,0 @@
-exports.Cli            = require('./lib/src').Cli;
-exports.CliRouter      = require('./lib/src/cli_router.js').CliRouter;
-exports.CliController  = require('./lib/src/cli_controller.js').CliController;
-exports.CliControllers = require('./lib/src/controllers');