+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+This project adheres to [Semantic Versioning](http://semver.org/).
+
 ## dev
 
-  * [TAG_NAME] change comment
+* Deprecations
+  * [api] Renaming CliController to Controller
+  * [api] Renaming CliControllers to Controllers
+  * [api] Moving export HelpController from CliControllers to HelpController
+
+* Enhancements
+  * [examples] Updating basic example
+  * [code] Replacing azk-dev por pure Babel
+  * [code] Replacing gulp for npm/yarn tasks